requirements.txt
mcp
fastapi
uvicorn
httpx
google-cloud-secret-manager
sse-starlette
pydantic-settings
python-dotenv
